[refactor]: moved files in their new repository according with new structure of project and rename some classes.
