+++ examples/docx/example.docx
